Add files via upload
Student Performance

Aim of the Analysis:

The aim of this analysis is to investigate the factors influencing student academic performance, particularly how parental education, gender, and test preparation impact grades in math, reading, and writing. The goal is to identify patterns and insights that can help improve student outcomes and guide educational support strategies.